pop message IDs from bot.d before raise instead of in handler
This does have the potential to result in two API calls to fetch a message, but it's only in cases where the user is both NOT a mod and attempting to un-sauce a Boxbot non-reply from another user, which isn't even possible right now I think. This avoids the datastore getting polluted in cases where two exceptions are thrown and the message can't be passed back out with the exception because it gets re-thrown by Lightbulb.