Add pixiv and oh god oh fuck I didn't commit at all during the creation of this ladle
Pixiv has been the source of a number of very annoying edge-cases in how I envision boxbot's new logic. Images can't be hotlinked, so I download the data and send it as an attachment. The new approach of using Resourceish types facilitates this nicely. Video on the other hand is very tricky, and it isn't clear how it fits into the existing response logic. I've put them in the main post with the embed, since otherwise the unsauce method just breaks (it's expecting multiple images and a specific field in the first post). It's nicer like this anyways.